Remove incorrect instruction to run 'yum update'
yum update installs available updates for all packages on the
system (not what we intend to advise here).  In reference to the
comment 'add instructions for yum update' removed in 5038c710:
this step is not necessary for yum, which will pull from the
available repositories during the next step (yum install).