Altered misleading sentence.
Create the ``/etc/apt/sources.list.d/10gen.list`` file and include
the following line for the 10gen repository.

>>

Create the ``/etc/apt/sources.list.d/10gen.list`` file and include
a line for the 10gen repository with the following command:

The original wording suggests that the following line should be added to the .list file, whereas the following line is actually a command to generate the .list file.