Updates serverStatus.mem.virtual
Since journalling is on by default, rewording the serverStatus.mem.virtual section is needed.

Removed the term "gigabytes", as this is subjective - a mem-map of 1MB shouldn't have 1GB as Virtual memory... But for 200GB, 1GB wouldn't be anything to worry about. A ratio of (for example) thrice mem-map, is better.