Update getting-started.txt
Inserted a comma on line 19.

Inserted a comma on line 30.

Inserted a comma on line 54.

Inserted a comma on line 55.

Inserted 'shell' on line 56.

Inserted a comma on line 66.

Inserted a comma on line 72.

Inserted a comma on line 81.

Inserted two commas on line 99.

Inserted a comma on line 129.

Inserted a comma on line 136.

Inserted a comma on line 152.

Inserted a comma on line 162.

Inserted a comma on line 174.

Inserted 'a' before 'query' on line 175.

On line 175, merged the second sentence with the first.

Inserted a comma on line 201.

Inserted a comma on line 243.

On line 278, replaced 'MongoDB JavaScript' with 'mongo'.

On lines 296 and 297, replaced 'although if the :doc:`ObjectId </reference/object-id>` values will be unique' with a new sentence.

On line 385, replaced 'The' with 'Your'.

Removed a superfluous 'only' from line 429.

Reordered the phrases on line 487 to make the sentence easier to read.

Signed-off-by: Sam Kleinman <samk@10gen.com>