Update fact-upgrade-sharded-cluster-prereq.rst
Unless I'm missing something here, the issue we identified with v:0 indexes isn't related to upgrade at all, and is no more or less likely after the upgrade.  We definitely want to encourage index upgrade, but not doing so won't make things worse.

Mention this because seeing customer concern about this warning in the docs.