Update fsync.txt
fsync+lock is not a shard-level operation, but a node level operation.

Additionally, it's critical to explain that fsync+lock is intended only for making file-system-level backups, and that
reads are not guaranteed not to block. (So that people don't
try to use this in conjunction with mongodump, which isn't
guaranteed to work, and sometimes hangs on people.)

Signed-off-by: Sam Kleinman <samk@10gen.com>