Refine "Combine Upsert/Multi" example
The example looked like a mistake because it resulted in the insertion
of a books document that only describes an author and not any actual
book. The collection being named "books", this seems counter
intuitive.

Hosting basic author information in the same database for an
application that has no dedicated author collection could make sense,
but the initial collection didn't seem to be intended for that.

Maybe this example should be changed entirely, but for now I've
extended the initial collection to show existing entries that match
this kind of expectation already.

The local author information would function as a cache of some sorts.