cloneCollection is not a viable alternative to copyTo
cloneCollection requires that the source and destination collections are on different mongods.  copyTo only works to copy a collection within a database on a single mongod.

Therefore, it is not possible to use cloneCollection instead of copyTo.  As such, we should not suggest the use of cloneCollection where copyTo does not behave as desired, since cloneCollection cannot work for that purpose and it is misleading to suggest that it can.