Update getting-started.txt
I added the word 'shell' after 'mongo' on line 28 for consistency. Everywhere else in the document where the mongo command line is referenced it is referred to as the 'mongo shell', but on line 28 it is being referred to as simply 'mongo'. This makes it difficult to understand without rereading it.

I added a comma on line 26 for increased readability.