DOCS-5129 Update CONTRIBUTING.rst
Correct the link for the contributor agreement. Previously,
the reference text was different from the intended text to be
linked. The reference text was not updated when 10gen
changed its name to MongoDB. The link in the README.rst
is correct, but when submitting a pull request, users are
directed by GitHub to the CONTRIBUTING.rst file by default.