Chunks can still migrate from a maxSize shard
It is untrue that a shard which is over its maxSize value will *not* have chunks migrated away from it.

This is correctly stated in the docs for the maxSize parameter for [addShard](http://docs.mongodb.org/manual/reference/command/addShard/), which says:

> The maxSize constraint prevents the balancer from migrating chunks to the shard when the value of mem.mapped exceeds the value of maxSize.

It is also easy to check empirically: start a 2 shard cluster, shard a collection, add `maxSize: 1` to the primary shard, then split once so there are 2 chunks, and observe that the balancer still migrates a chunk away from the primary (so that it's now 1 chunk on each shard).  If you manually move both chunks onto the *non-maxSize* shard, the balancer will not move a chunk onto the maxSize shard (which is correct).