DOCS-6874 fix count with read concern
The readConcern() shell helper does not work with db.collection.count(), as
the latter is not a cursor method. (It returns an integer.)

Instead, use the options parameter with db.collection.count() to specify a
read concern. You can also chain cursor.count() to a db.collection.find()
command with a read concern, like so:

    db.collection.find().readConcern(<read-concern>).count()