minor: read concern fixes
When writing in text, favor "read concern." The camel-cased "readConcern" is
acceptable in code or when quoted in backticks as a literal. The read concern
and write concern pages now match each other in terms of naming.

This commit also has some minor formatting and grammar fixes in various read
concern-related pages.

Signed-off-by: kay <kay.kim@10gen.com>