Index intersection and sort
I cannot follow the document as it is, so I wonder whether there was a mistake in it. You say that query intersection cannot be used for the query, and that individual indexes will also not be used. In other words, can the query not be optimised at all? That sounds unlikely. Did you mean what I am suggesting here?