Corrected ordered duplicate key thrown exception
The thrown exception did not match what is explained in the text above and below it.
It was showing an error while inserting the first document (_id:13), but actually the error is thrown while inserting the second document (_id:13 again). So, "op" and "nInserted" were corrected, as well as the database and collection names in the "errmsg" to be consistent with the other examples in the same documentation.