Added note about = sign for archive filename
It is not clear form the option docs that you have to supply `--archive=<filename>` while most other options can be supplied without the `=` sign.  Although the examples at the bottom of the page do show the = sign, it is nowhere mentioned in the docs it is required which was quite confusing. This commit should better explain that