Include CA Private key in test-ca.pem
Following these instructions exactly with the latest version of Mongo yields an error on mongod restart.  The error would be with the test-ca.pem file stating "error:no start line".  OpenSSL is expecting a private key at the beginning of the file.  Adding the private key as the first element to concatenate to the test-ca.pem file resolves the issue.