update document
It seems to me that "Update  Values in an Array" phase means $ postional operator changes all values inside array. "Update  a Value in an Array" is more close to what $ is going todo. 
The same is about  Update Documents in an Array. It would be more close what $ positional operator is going to do if we change it to "Update a Document in an Array".