DOP-3026: publish manifests to correct bucket/path (#1368)
Search is now using the docs-search-indexes-test bucket, not the old bucket(s) we were using previously. This updates the Makefile that y'all use for deploys to put the manifests into the right bucket.

I've manually copied the existing manifests over so that search will work as designed from the main docs search, this will just ensure that updates end up in the right place. You'll probably want to back-port the changes to the earlier editions that are actively maintained (if indeed there are any?).