MONGOID-5228 disallow _id to be updated on persisted documents (#5542)
* MONGOID-5228 - disallow _id to be updated on persisted documents

Previously, updating _id on any document (top-level or embedded) would silently ignore the _id attribute, making it appear as though the operation succeeded. This was misleading, since the _id didn't actually change. This change makes it so that _id is explicitly made immutable once a document has been persisted. The same behavior is intentionally enforced regardless of whether the document is embedded or not. If you find yourself needing to update the _id field in an embedded document, you'll need to use the lower-level driver methods to accomplish this. We believe modifying the _id field (even in embedded documents) is an anti-pattern.

* use a specialized error for immutability violations

* add a feature flag

* change the flag name, and account for existing nested attribute behavior

* wrong default for <9.0

* clean up a test that was unnecessarily sensitive to the _id changing

* another test that was unnecessarily sensitive to _id changes

* more tests that are too sensitive to _id changes

* we need to allow _id mutations if the _id was previously nil

* mention enforcement of _id immutability in the release notes