Hotfix for table with incorrect wrapping (#6)
Follow-up PR to the merged migration PR: the new-gen rendering of the
table on this page is not ideal because monospaced text excerpts won't
wrap. This PR fixes that. The screenshot on this PR shows the "before"

[Staging](https://preview-mongodbsarahsimpers.gatsbyjs.io/ops-manager/hotfix-34607-2/tutorial/manage-s3-blockstore-storage/#procedures)

[Build](https://workerpool-boxgs.mongodbstitch.com/pages/job.html?collName=queue&jobId=65805dad7a5cb5fef542006a)
<img width="788" alt="Screenshot 2023-12-17 at 1 25 26 PM"
src="https://github.com/10gen/docs-ops-manager/assets/82042374/a7867576-cb0f-41a5-b241-54b11131d797">