Removes ops manager from the intersphinx list, term URL syntax for new gen (#15)
[Jira](https://jira.mongodb.org/browse/DOCSP-35182)

[Staging](https://preview-mongodbsarahsimpers.gatsbyjs.io/ops-manager/update-intersphinx/)

[Build](https://workerpool-boxgs.mongodbstitch.com/pages/job.html?collName=queue&jobId=6584476a77a86a479bb6f296)

Per JW's advice on the migration PR, removing Ops Manager from the
intersphinx list to avoid errors at publication time. This PR also
resolves the errors across about 30 pages that arose after removing it
from the intersphinx. These changes are in the first commit.

Aside from that, I noticed another issue: in new gen docs, a link with
#term-abcde doesn't work. It has to be #std-term-abcde. That means that
CURRENTLY #term-abcde points to the right place in the old gen docs, but
won't when we migrate. So I changed all opsmgr glossary term references
to be std-term via regex find/replace. This means the links won't work
now (will only bring you to the glossary page, not the term itself on
old-gen), but will work once it's on new gen. I did this in commit 2.