(DOCSP-46035) Rotate LDAP pwd for Agent in Cloud Manager (#150)
This PR adds a new procedure for rotating an LDAP password for the
Agent, in Cloud Manager.
It reuses an existing procedure for setting up LDAP, and adds a new
field, New Query Password.
All other changes in this PR are edits not related to this feature. For
example, I changed :doc: type refs to :ref: type references in a few
places. I also streamlined text in a few places and removed refs to
MongoDB 3.4 because it hasn't been supported for a while and is not
worth mentioning.

@pstefek-mongo please review.

- [DOCSP-46035](https://jira.mongodb.org/browse/DOCSP-46035)
- [Manage LDAP Authentication for your Cloud Manager
Project](https://deploy-preview-150--docs-cloud-manager.netlify.app/tutorial/enable-ldap-authentication-for-group/)
-- this is the top-most topic in the docs
- [Rotate LDAP Authentication
Passwords](https://deploy-preview-150--docs-cloud-manager.netlify.app/tutorial/enable-ldap-authentication-for-group/#rotate-ldap-authentication-passwords)
- this is the section for rotating that is new (new steps, modeled after
existing steps).