(DOCSP-45035) Rotate LDAP pwd for Agent in OM (#675)
This PR is equivalent in its contents to the [PR on the Cloud Manager
side](https://github.com/10gen/docs-cloud-manager/pull/150). Note: These
doc sources are now split and hence we have this separate PR, as opposed
to having a single PR for OM/CM docs that were previously built from one
source.

The only difference is that I reused the same steps file for both
enabling LDAP and rotating the pwd. This is because OM uses yaml file
includes and it was easier to do it this way for this long-maintained
doc repo. The rest of the approach is the same as for the CM PR.

@pstefek-mongo and @kyuan-mongodb could you please review? 

- [DOCSP-45035](https://jira.mongodb.org/browse/DOCSP-45035)
-
[STAGING](https://deploy-preview-675--docs-ops-manager.netlify.app/tutorial/enable-ldap-authentication-for-group/)