(DOCSP-49806) Clarify misleading parameters description. (#858)
The [current
doc](https://www.mongodb.com/docs/ops-manager/current/core/api/)
explains the "parameters" field in error as:

"List of parameters included in the Atlas Administration API request"

This is quite misleading since this list is returning additional
information for the error (no correlation with the request params).

In the Java this field is
[documented](https://github.com/10gen/mms/blob/d1915c2f3ad23a9707f5a66418390d60807936f9/server/src/main/com/xgen/svc/mms/api/res/common/ApiError.java#L62)
as: "Parameters used to give more information about the error.". Which
seems to me worded much better and should be used in the public
documents as is.
- [DOCSP-49806](https://jira.mongodb.org/browse/DOCSP-49806)
-
[STAGING](https://deploy-preview-858--docs-ops-manager.netlify.app/core/api/#errors)

### Self-Review Checklist

- [ ]
[Define](https://wiki.corp.mongodb.com/display/DE/Taxonomy+tagging+instructions)
taxonomy
[values](https://wiki.corp.mongodb.com/display/DE/Docs+Taxonomy) at top
of page.
- [ ] Add genre facets (tutorial or reference), as in this [example
PR](https://github.com/10gen/cloud-docs/pull/5042).
- [ ] Add programmingLanguage (if necessary).
- [ ] Add meta keywords (if necessary).
- [ ] Resolve any new warnings or errors in the build.
- [ ] Proofread for spelling and grammatical errors.
- [ ] Check staging for rendering issues.
- [ ] Confirm links are working.