DOCSP-55178-unique-indexes-mini-rework (#16054)
## Description

Promote usage of `prepareUnique` by highlighting it in the unique
indexes landing page. As part of these changes I did a light rework of
the unique indexes content to better align with our content templates.

There is probably more that can be done to improve the organization of
this content, but I think this is a step in the right direction.

## Staging

- [Unique
Indexes](https://deploy-preview-16054--mongodb-manual.netlify.app/docs/upcoming/core/index-unique/)
(landing)
-
[Single-Field](https://deploy-preview-16054--mongodb-manual.netlify.app/docs/upcoming/core/index-unique/create/)
-
[Compound](https://deploy-preview-16054--mongodb-manual.netlify.app/docs/upcoming/core/index-unique/create-compound/#create-a-unique-compound-index)

## JIRA

https://jira.mongodb.org/browse/DOCSP-55178