DOP-6328:  NotFound Template (#16266)
### TICKET
[DOP-6328](https://jira.mongodb.org/browse/DOP-6328) Next.js: NotFound
Template

### NOTES
Was already implemented by @caesarbell . This pr just removes old
comments ported over from Snooty that are no longer necessary. Instead
of implementing the not-found page as a template, we leverage Next.js
default functionality; it's automatically recognized as the 404 handler
for that route segment. When notFound() is called or no route matches,
Next.js renders the nearest not-found.tsx file in the route hierarchy.

ex:
https://deploy-preview-16266--docs-on-nextjs.netlify.app/docs/abcd/