DOCSP-50202: Make it clear in the public documents that metadata needs to be manually cleaned after successful migration (#16365)
DOCSP-50202

Adds information about dropping metadata to quickstart. I put this in
the "finalize cutover process" section, but i'm not sure if this is the
best place for this, or if it should go under notes or in the "finalize
cutover process" page itself.

Additionally, adds information here about not being able to reverse
migration after metadata has been dropped.

-
https://deploy-preview-16365--mongodb-mongosync.netlify.app/docs/mongosync/current/quickstart/#finalize-cutover-process
-
https://deploy-preview-16365--mongodb-mongosync.netlify.app/docs/mongosync/current/reference/mongosync/mongosync-behavior/#mongosync-metadata