DOCSP-55720 C November Repo Steward Tasks (#16930)
Fixes build errors in the C driver docs

Build log -
https://app.netlify.com/projects/mongodb-c-driver/deploys/69459a3ba5deb100082f502c
Jira - https://jira.mongodb.org/browse/DOCSP-55720

Note: there is a recurring error of:
```
The "poetry.dev-dependencies" section is deprecated and will be removed in a future version. Use "poetry.group.dev.dependencies" instead.
The currently activated Python version 3.14.2 is not supported by the project (>=3.8,<3.13).
Trying to find and use a compatible version.
Using python3.12 (3.12.3)
```

I don't think this is in the C driver docs, I searched
"poetry.dev-dependencies" in the docs and got zero results. I think this
is something on the configuration side and can be ignored.


There was also another recurring error in the older docs versions about
the CMakeLists file, but I think that can be disregarded because the
parser seems to be reading the file as a docs file when it is a C
configuration file and therefore the formatting cannot be changed to fit
what the parser expects.